Move Stack operations from Stack to StackActions
